Updated memory addresses, poker, in-progress custom res support
Removed a lot of bloat in the UI. Began rewriting the ultrawide
support to work more efficiently. Moved poker methods from variables
to one-liners for methods only requiring one memory access and using
arrays and for-loops for methods that require more than one. Updated
all of the memory addresses for the latest version of Halo 5: Forge.
Will eventually move the addresses over to pattern/mask scanning.